api/requirements.txt
fastapi
weaviate-client
neo4j
python-dotenv
uvicorn
